Update Dockerfile for Debian Bullseye release
The previously used base image `python:3.7` was changed to now use
Debian Bullseye. This lead to installation issues of `libenchant-dev`
(see #31).

This is now fixed by upgrading the installation command for `libenchant`
to `libenchant-2-dev` (see also https://pkgs.org/search/?q=LIBENCHANT).

Additionally, to avoid similar issues in the future, the Debian version
of the base image is now defined in the `Dockerfile`.